\name{print8pam}
\alias{print8pam}
\title{
prints a /pam/
}
\description{
(ba) Nothing more than printing a /pam/ matrix (after checking it).
}
\usage{
print8pam(x, ...)
}
\examples{
 g4n3k("RESET"); # needed only for R checking, to be forgotten
 print(g4n.pam0);
}
\arguments{
\item{x}{pam object}
\item{\dots}{Further arguments to be passed to the print function.}
}
\value{
nothing a printing is issued
}
\keyword{print}
