\name{valid8arc}
\alias{valid8arc}
\title{
checks a /arc/
}
\description{
(ba) This function checks /arc/ objects
}
\usage{
valid8arc(object)
}
\arguments{
\item{object}{The arc object to be validated.}
}
\details{
It is the validity method for /arc/ objects.
}
\value{
TRUE when the object seems acceptable else a character describing the
error(s)
}
\keyword{classes}
