\name{valid8gn}
\alias{valid8gn}
\title{
checks a /gn/
}
\description{
This function checks /gn/ objects
}
\usage{
valid8gn(object)
}
\arguments{
\item{object}{The gn object to be validated.}
}
\details{
It is the validity method for /gn/ objects.
}
\value{
TRUE when the object seems acceptable else a character describing the
error(s)
}
\keyword{classes}
