\name{valid8pos}
\alias{valid8pos}
\title{
checks a /pos/
}
\description{
(ba) This function checks /pos/ objects
}
\usage{
valid8pos(object)
}
\arguments{
\item{object}{The pos object to be validated.}
}
\details{
It is the validity method for /pos/ objects.
}
\value{
TRUE when the object seems acceptable else a character describing the
error(s)
}
\keyword{classes}
